hatch-455: fix default includes for Everything
