+++ packages/react-ui/src/components/HatchifyCollection/index.ts
@@ -1,2 +1,5 @@
 export type { HatchifyCollectionProps } from "./HatchifyCollection"
-export { default as HatchifyCollection } from "./HatchifyCollection"
+export {
+  default as HatchifyCollection,
+  getDefaultInclude,
+} from "./HatchifyCollection"
