+++ packages/react-ui/src/components/HatchifyEverything/HatchifyEverything.tsx
@@ -1,4 +1,3 @@
-import { useState } from "react"
 import type {
   Filters,
   PaginationObject,
@@ -6,10 +5,8 @@ import type {
 } from "@hatchifyjs/rest-client"
 import type { PartialSchema } from "@hatchifyjs/core"
 import type { HatchifyReactRest } from "@hatchifyjs/react-rest"
-import { useHatchifyPresentation } from ".."
 import type { HatchifyCollectionSelected, SortObject } from "../../presentation"
-import useCollectionState from "../../hooks/useCollectionState"
-import type { GetSchemaNames } from "@hatchifyjs/rest-client"
+import { WithSchemas, NoSchemas } from "./components"
 
 export interface HatchifyEverythingProps<
   TSchemas extends Record<string, PartialSchema>,
@@ -31,60 +28,11 @@ function HatchifyEverything<
   return (
     <>
       {Object.keys(finalSchemas).length !== 0 ? (
-        <HatchifyEverythingWithSchema {...rest} finalSchemas={finalSchemas} />
+        <WithSchemas {...rest} finalSchemas={finalSchemas} />
       ) : (
-        <HatchifyEverythingNoSchema />
+        <NoSchemas />
       )}
     </>
-  )
-}
-
-function HatchifyEverythingNoSchema(): JSX.Element {
-  const { Everything } = useHatchifyPresentation()
-
-  return <Everything />
-}
-
-function HatchifyEverythingWithSchema<
-  const TSchemas extends Record<string, PartialSchema>,
->({
-  finalSchemas,
-  partialSchemas,
-  restClient,
-  children,
-  defaultSelected,
-  onSelectedChange,
-  defaultPage,
-  defaultSort,
-  baseFilter,
-}: HatchifyEverythingProps<TSchemas>): JSX.Element {
-  const schemasList = Object.keys(finalSchemas) as Array<
-    GetSchemaNames<TSchemas>
-  >
-
-  const [selectedSchema, setSelectedSchema] = useState(schemasList[0])
-  const { Everything } = useHatchifyPresentation()
-
-  const collectionState = useCollectionState(
-    finalSchemas,
-    partialSchemas,
-    selectedSchema,
-    restClient,
-    {
-      defaultSelected,
-      onSelectedChange,
-      defaultPage,
-      defaultSort,
-      baseFilter,
-    },
-  )
-
-  return (
-    <Everything
-      {...collectionState}
-      schemaName={selectedSchema as string}
-      setSelectedSchema={setSelectedSchema}
-    />
   )
 }
