+++ packages/react-ui/src/components/HatchifyEverything/components/NoSchemas.tsx
@@ -1,0 +1,7 @@
+import { useHatchifyPresentation } from "../../HatchifyPresentationProvider"
+
+export function NoSchemas(): JSX.Element {
+  const { Everything } = useHatchifyPresentation()
+
+  return <Everything />
+}
